Add metadata extraction API for reprostim-videocapture logs and create new capture.metadata module
